0711 - 검색 내역 구현, ModalSearchController 개선
