+++ src/js/modal/ModalSearchController.js
@@ -11,22 +11,35 @@ export default class ModalSearchController {
         this.searchHistoryAry = []
 
         this.bindEvents();
+        this.setState()
+    }
+
+    addSearchHistory(searchKeyword) {
+        if (this.searchHistoryAry !== null){
+            const historyPos = this.searchHistoryAry.indexOf(searchKeyword)
+            if (historyPos >= 0){
+                this.searchHistoryAry.splice(historyPos, 1)
+            }
+            if (this.searchHistoryAry.length > 2) {
+                this.searchHistoryAry.shift()
+            }
+        }
+        this.searchHistoryAry.push(searchKeyword)
+        saveDataToLocalStorage("searchHistory", this.searchHistoryAry)
+        this.setState()
     }
 
     onSearchFormSubmit(event){
+        const searchKeyword = this.$searchInput.value
         event.preventDefault()
-        this.onSubmit(this.$searchInput.value)
+        this.addSearchHistory(searchKeyword)
+        this.onSubmit(searchKeyword)
     }
 
     onSearchHistoryClick(event){
         const searchKeyword = event.target.innerHTML
-        const historyPos = this.searchHistoryAry.indexOf(searchKeyword)
-        if (historyPos >= 0){
-            this.searchHistoryAry.splice(historyPos, 1)
-        }
-        this.searchHistoryAry.push(searchKeyword)
+        this.addSearchHistory(searchKeyword)
         this.onSubmit(searchKeyword)
-        console.log(this.searchHistoryAry)
     }
 
     bindEvents(){
@@ -34,4 +47,32 @@ export default class ModalSearchController {
         this.$searchButton.addEventListener("click", (event) => this.onSearchFormSubmit(event))
         this.$seacrhHistory.addEventListener("click", (event) => this.onSearchHistoryClick(event))
     }
+
+    render(){
+        this.$seacrhHistory.innerHTML = buildSearchHistorySection(this.searchHistoryAry)
+    }
+
+    setState(){
+        this.searchHistoryAry = loadDataFromLocalStorage("searchHistory")
+        this.render()
+    }
+}
+
+const buildSearchHistorySection = (searchHistory) => {
+    let searchHistoryDom = `<span class="text-gray-700">최근 검색어: </span>`
+    for(let i = searchHistory.length - 1; i >=0; i--){
+        searchHistoryDom += `<a class="chip">${searchHistory[i]}</a>`
+    }
+    return searchHistoryDom
+}
+
+const saveDataToLocalStorage = (key, value) => {
+    const jsonValue = JSON.stringify(value)
+    localStorage.setItem(key, jsonValue)
+}
+
+const loadDataFromLocalStorage = (key) => {
+    const searchHistory = JSON.parse(localStorage.getItem(key))
+    if (searchHistory === null) return []
+    else return searchHistory
 }
